chore(clients): split invoke timeout helper to satisfy SwiftLint length cap
